Move filter_gen tests into individual files
